+++ composition-js/src/__tests__/supergraph_reversibility.test.ts
@@ -1,7 +1,6 @@
 import { assertCompositionSuccess, composeAsFed2Subgraphs } from "./testHelper";
 import gql from 'graphql-tag';
-import { asFed2SubgraphDocument, buildSubgraph, FEDERATION2_LINK_WITH_AUTO_EXPANDED_IMPORTS, ServiceDefinition, Supergraph } from "@apollo/federation-internals";
-import { composeServices } from "../compose";
+import { asFed2SubgraphDocument, buildSubgraph, ServiceDefinition, Supergraph } from "@apollo/federation-internals";
 
 function composeAndTestReversibility(subgraphs: ServiceDefinition[]) {
   const result = composeAsFed2Subgraphs(subgraphs);
@@ -128,170 +127,5 @@ describe('@interfaceObject', () => {
     };
 
     composeAndTestReversibility([s1, s2]);
-  });
-});
-
-
-describe('demand control directives', () => {
-  it('extracts @cost from the supergraph', () => {
-    const subgraph = {
-      name: 'my-subgraph', 
-      typeDefs: asFed2SubgraphDocument(gql`
-        extend schema @link(url: "https://specs.apollo.dev/cost/v0.1", import: ["@cost"])
-
-        type Query {
-          fieldWithCost: Int @cost(weight: "5.0")
-        }
-      `)
-    };
-
-    const result = composeServices([subgraph]);
-    assertCompositionSuccess(result);
-    expect(result.hints.length).toBe(0);
-    const extracted = Supergraph.build(result.supergraphSdl).subgraphs().get('my-subgraph');
-
-    expect(extracted?.toString()).toMatchString(`
-      schema
-        ${FEDERATION2_LINK_WITH_AUTO_EXPANDED_IMPORTS}
-      {
-        query: Query
-      }
-
-      type Query {
-        fieldWithCost: Int @federation__cost(weight: "5.0")
-      }
-    `);
-  });
-
-  it('extracts @listSize from the supergraph', () => {
-    const subgraph = {
-      name: 'my-subgraph', 
-      typeDefs: asFed2SubgraphDocument(gql`
-        extend schema @link(url: "https://specs.apollo.dev/listSize/v0.1", import: ["@listSize"])
-
-        type Query {
-          bigList: [String!] @listSize(assumedSize: 2000, requireOneSlicingArgument: false)
-        }
-      `)
-    };
-
-    const result = composeServices([subgraph]);
-    assertCompositionSuccess(result);
-    expect(result.hints.length).toBe(0);
-    const extracted = Supergraph.build(result.supergraphSdl).subgraphs().get('my-subgraph');
-
-    expect(extracted?.toString()).toMatchString(`
-      schema
-        ${FEDERATION2_LINK_WITH_AUTO_EXPANDED_IMPORTS}
-      {
-        query: Query
-      }
-
-      type Query {
-        bigList: [String!] @federation__listSize(assumedSize: 2000, requireOneSlicingArgument: false)
-      }
-    `);
-  });
-
-  it('extracts the correct @cost for different subgraphs with @shareable fields', () => {
-    const subgraphA = {
-      name: 'subgraph-a',
-      typeDefs: asFed2SubgraphDocument(gql`
-        extend schema @link(url: "https://specs.apollo.dev/cost/v0.1", import: ["@cost"])
-
-        type Query {
-          sharedWithCost: Int @shareable @cost(weight: "5.0")
-        }
-      `)
-    };
-    const subgraphB = {
-      name: 'subgraph-b',
-      typeDefs: asFed2SubgraphDocument(gql`
-        extend schema @link(url: "https://specs.apollo.dev/cost/v0.1", import: ["@cost"])
-
-        type Query {
-          sharedWithCost: Int @shareable @cost(weight: "10.0")
-        }
-      `)
-    };
-
-    const result = composeServices([subgraphA, subgraphB]);
-    assertCompositionSuccess(result);
-    expect(result.hints.length).toBe(0);
-    const supergraph = Supergraph.build(result.supergraphSdl);
-
-    expect(supergraph.subgraphs().get(subgraphA.name)?.toString()).toMatchString(`
-      schema
-        ${FEDERATION2_LINK_WITH_AUTO_EXPANDED_IMPORTS}
-      {
-        query: Query
-      }
-
-      type Query {
-        sharedWithCost: Int @shareable @federation__cost(weight: "5.0")
-      }
-    `);
-    expect(supergraph.subgraphs().get(subgraphB.name)?.toString()).toMatchString(`
-      schema
-        ${FEDERATION2_LINK_WITH_AUTO_EXPANDED_IMPORTS}
-      {
-        query: Query
-      }
-
-      type Query {
-        sharedWithCost: Int @shareable @federation__cost(weight: "10.0")
-      }
-    `);
-  });
-
-  it('extracts the correct @listSize for different subgraphs with @shareable fields', () => {
-    const subgraphA = {
-      name: 'subgraph-a',
-      typeDefs: asFed2SubgraphDocument(gql`
-        extend schema @link(url: "https://specs.apollo.dev/listSize/v0.1", import: ["@listSize"])
-
-        type Query {
-          sharedWithListSize: [Int] @shareable @listSize(assumedSize: 10)
-        }
-      `)
-    };
-    const subgraphB = {
-      name: 'subgraph-b',
-      typeDefs: asFed2SubgraphDocument(gql`
-        extend schema @link(url: "https://specs.apollo.dev/listSize/v0.1", import: ["@listSize"])
-
-        type Query {
-          sharedWithListSize: [Int] @shareable @listSize(assumedSize: 20)
-        }
-      `)
-    };
-
-    const result = composeServices([subgraphA, subgraphB]);
-    assertCompositionSuccess(result);
-    expect(result.hints.length).toBe(0);
-    const supergraph = Supergraph.build(result.supergraphSdl);
-
-    expect(supergraph.subgraphs().get(subgraphA.name)?.toString()).toMatchString(`
-      schema
-        ${FEDERATION2_LINK_WITH_AUTO_EXPANDED_IMPORTS}
-      {
-        query: Query
-      }
-
-      type Query {
-        sharedWithListSize: [Int] @shareable @federation__listSize(assumedSize: 10)
-      }
-    `);
-    expect(supergraph.subgraphs().get(subgraphB.name)?.toString()).toMatchString(`
-      schema
-        ${FEDERATION2_LINK_WITH_AUTO_EXPANDED_IMPORTS}
-      {
-        query: Query
-      }
-
-      type Query {
-        sharedWithListSize: [Int] @shareable @federation__listSize(assumedSize: 20)
-      }
-    `);
   });
 });
